% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CEOcomp}
\alias{CEOcomp}
\title{Compensation of CEOs and firm characteristics}
\format{A data frame with 30 rows and 14 variables:
\itemize{
  \item compvar: variable compensation [kEUR]
  \item comptotal: total compensation [kEUR]
  \item roa: return on assets [percent]
  \item roe: return on equity [percent]
  \item tsr: total shareholder return [percent]
  \item emp: employees [number of employees]
  \item assets: total assets [kEUR]
  \item beta: company's beta (risk measure)
  \item marketcap: market capitalization [mEUR]
  \item age: CEO's age [years]
  \item tenure: tenure for CEO position [years]
  \item listed: dummy (factor) indicating wheter company is listed [0 and 1]
}}
\usage{
CEOcomp
}
\description{
A dataset containing the compensation of CEOs of 30 firms along with
 some firm characteristics. The data are fictional but realistic.
 The variables are as follows:
}
\keyword{datasets}

