% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_all_formulae.R
\name{generiere_alle_formeln}
\alias{generiere_alle_formeln}
\title{Generate all formulae from given inputs}
\usage{
generiere_alle_formeln(abhVar, opList, festVar)
}
\arguments{
\item{abhVar}{A todo.}

\item{opList}{A todo.}

\item{festVar}{todo.}
}
\value{
all formulae as list of formula.
}
\description{
Generate all formulae from given inputs
}

