% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sub_vars_rhs_one2one}
\alias{sub_vars_rhs_one2one}
\title{sub_vars_rhs_one2one - Positionstreue ist ein Feature}
\usage{
sub_vars_rhs_one2one(form, old, new)
}
\arguments{
\item{form}{formula}

\item{old}{zu ersetzende Variablen - stringvector}

\item{new}{neue Variablen - stringvector}
}
\value{
formula aktualisierte Formel
}
\description{
sub_vars_rhs_one2one - Positionstreue ist ein Feature
}

